requirements.txt
torch
transformers
peft
python-dotenv
wandb
datasets
omegaconf
lm_eval
sentencepiece
jupyter
transformer-lens
matplotlib
Levenshtein
openai
fastchat
deepspeed
pandas
numpy
